refactors testing layout
