+++ app/src/test/java/Bamboo/controllerTesting/SettingsTest.java
@@ -1,4 +1,4 @@
-package Bamboo;
+package Bamboo.controllerTesting;
 
 import Bamboo.controller.Agent;
 import Bamboo.controller.Settings;
