+++ app/src/test/java/Bamboo/minimaxTesting/MiniMaxNodeTreeTest.java
@@ -1,4 +1,4 @@
-package Bamboo;
+package Bamboo.minimaxTesting;
 
 import Bamboo.controller.miniMax.NodeMM;
 import Bamboo.model.*;
